pyproject.toml
[tool.poetry]
name = "psycopg3-example-poetry"
version = "0.1.0"
description = ""
authors = ["j0no"]

[tool.poetry.dependencies]
python = "^3.10"
psycopg = {extras = ["binary"], version = "^3.0.11"}
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
